Add securityContext `allowPrivilegeEscalation: false` config for runner pod containers
Set allowPrivilegeEscalation to false all the time (hardcoded) and we don't allow the user to change this since there shouldn't be a need.
The gitlab-runner process doesn't do this at all, since this setting is for the deployment of the gitlab-runner the jobs run in a separate pod with their own settings.

Closes #183